requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.2
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
zope.interface==6.0